refactor: Refactor code styles until not complete HelpSideBar
I change styled jsx to styled components, and divide tsx file by styled file.
#31
