+++ client/src/components/organisms/HelpSideBar.tsx
@@ -1,13 +1,14 @@
 import { Box } from '@mui/system';
 import { connect } from 'react-redux';
 import { closeHelpSideBar, RootState } from '~/others/store';
-import { HelpCallBox, HelpFinBox } from '../molecules/HelpCallBox';
+import { HelpCallBox, HelpFinBox } from '../molecules/HelpBoxes.tsx';
+import styled from 'styled-components';
 
 interface HelpSideBarProps {
-  state: RootState;
+  isHelpSideBarOpen: boolean;
 }
 
-const HelpSideBar = ({ state }: HelpSideBarProps) => {
+const HelpSideBar: React.FC<HelpSideBarProps> = ({ isHelpSideBarOpen }) => {
   return (
     <>
       <Box
@@ -26,33 +27,17 @@ const HelpSideBar = ({ state }: HelpSideBarProps) => {
         onClick={closeHelpSideBar}
       />
 
-      <div className='block'></div>
-      <div className='helpSideBar'>
+      <Block className='block'></Block>
+      <StyledHelpSideBar className='helpSideBar'>
         <HelpCallBox />
         <HelpCallBox />
         <HelpCallBox />
         <HelpFinBox />
         <HelpFinBox />
         <HelpFinBox />
-      </div>
+      </StyledHelpSideBar>
       <style jsx>{`
-        .block {
-          width: 300px !important;
-          min-width: 300px;
-          height: calc(100vh - 70px);
-        }
-        .helpSideBar {
-          position: fixed;
-          top: 70px;
-          right: 0;
-          width: 300px !important;
-          min-width: 300px;
-          height: calc(100vh - 70px);
-          background: #fff;
-          border-left: solid 1px #ddd;
-          z-index: 2;
-        }
-        ${state.helpSideBarReducer
+        ${isHelpSideBarOpen
           ? ``
           : `
         .block, .hide, .helpSideBar{
@@ -65,8 +50,34 @@ const HelpSideBar = ({ state }: HelpSideBarProps) => {
   );
 };
 
+interface VisibleDivProps {
+  isHelpSideBarOpen: boolean;
+}
+
+const VisibleDiv = styled.div<VisibleDivProps>`
+  ${(props) => (props.isHelpSideBarOpen ? '' : 'display: none;')}
+`;
+
+const Block = styled.div`
+  width: 300px !important;
+  min-width: 300px;
+  height: calc(100vh - 70px);
+`;
+
+const StyledHelpSideBar = styled.div`
+  position: fixed;
+  top: 70px;
+  right: 0;
+  width: 300px !important;
+  min-width: 300px;
+  height: calc(100vh - 70px);
+  background: #fff;
+  border-left: solid 1px #ddd;
+  z-index: 2;
+`;
+
 const mapStateToProps = (state: RootState) => {
-  return { state };
+  return { isHelpSideBarOpen: state.helpSideBarReducer };
 };
 
 export default connect(mapStateToProps)(HelpSideBar);
